Inventory serialisation and from_dict. Full entity object graph coverage!
