+++ src/entities/item/equippable.py
@@ -230,7 +230,7 @@ class EquippableConfig(NamedTuple):
             "range": self.range,
             "attacks": self.attacks,
             "spells": self.spells,
-            "affixes": [affix.to_dict() for affix in self.affixes],
+            # "affixes": [affix.to_dict() for affix in self.affixes],
         }
